feat: revised LightAccount v2 1271 sigs
Updates the `signMessage` and `signTypedData` implementations for
`LightAccount` v2. The updated version no longer attempts to use nested
typed data to display the original message to the user, as doing so
creates a security risk in which valid signing requests can be
disguished in a way that concels their original contents.

Instead, the new `LightAccount` v2 uses the same signing mechanism as
v1.1.0, except that it also prepends an enum indicating if the owner is
an EOA or smart contract account, as a way to both save gas and to
provide more accurate gas estimates.