fix: close OAuth popup from parent (#973)
On iOS, a browser window can't close itself via `window.close()`:
instead, it must be closed by the same context that called `window.open`
to open it. This means that for iOS we cannot rely on the OAuth callback
window to close itself in the popup-based flow. Instead, we'll have the
parent close it upon receiving a message.