fix: open the auth modal upon email return
On mobile, if the user clicks an email magic link, the modal should open
immediately to display that they are completing email login.

This has the side effect of opening the modal for OAuth redirect login
too, because at the moment email and OAuth share a loading state when
returning.