feat(connectors-web): add alchemySmartWallet connector
Basic implementation of a connector which uses the Wallet API to perform
actions using an Alchemy smart account, wrapping another connector to
perform signing.

This version is tested to work when wrapping both the Metamask connector
and the Alchemy Auth connector, including many edge cases such as
changing accounts or configs in the Metamask UI after connecting.