feat(connectors-web): add alchemySmartWallet connector
Basic implementation of a connector which uses the Wallet API to perform
actions using an Alchemy smart account, wrapping another connector to
perform signing.

This version is tested to work with the Metamask connector, although
there is additional work that can done to clean up code and polish edge
cases, as marked with TODO's.