feat(connectors-web): add alchemySmartWallet connector (#2115)
* feat(connectors-web): add alchemySmartWallet connector

Basic implementation of a connector which uses the Wallet API to perform
actions using an Alchemy smart account, wrapping another connector to
perform signing.

This version is tested to work when wrapping both the Metamask connector
and the Alchemy Auth connector, including many edge cases such as
changing accounts or configs in the Metamask UI after connecting.

* feat: use smart wallet connector to send calls in connector web example (#2152)

* feat: use smart wallet connector to send calls in connector web example

* chore: remove bad import

* refactor: improve smart wallet 1193 provider (#2153)

* refactor: improve smart wallet 1193 provider

* chore: remove debug log

* refactor: add createSmartWalletClientAndRequestAccount helper

* chore: type import

* feat(v5): implement smartWalletConnector in react demo (#2158)

---------

Co-authored-by: jakehobbs <jacob.hobbs@alchemy.com>